tests/wpt/css-tests/css-transforms-1_dev/xhtml1/css-transform-property-existence.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <meta charset="utf-8" />
    <title>CSS Transforms Test: property existence</title>
    <link href="http://www.intel.com" rel="author" title="Intel" />
    <link href="http://www.w3.org/TR/css-transforms-1/#propdef-transform" rel="help" />
    <meta content="dom" name="flags" />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="test"></div>
    <div id="log"></div>
    <script>
      test(function() {
        assert_not_equals(document.getElementById("test").style.transfor, undefined, "expect transform is not undefined");
      }, "Check the existence of transform.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "transform-origin", "expect transform-origin");
      }, "Check the existence of transform-origin.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "transform-style", "expect transform-style");
      }, "Check the existence of transform-style.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "perspective", "expect perspective");
      }, "Check the existence of perspective.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "perspective-origin", "expect perspective-origin");
      }, "Check the existence of perspective-origin.");

      test(function() {
        assert_own_property(document.getElementById("test").style, "backface-visibility", "expect backface-visibility");
      }, "Check the existence of backface-visibility.");
    </script>
  

</body></html>
